requirements.txt
discord.py-self[voice]
PyNaCl
loguru
yt-dlp
pytubefix